Breaking: do not implicitly call parent constructors
Currently the parent constructors are always implicitly called with
final constructor arguments, which causes conflicts and unexpected
behaviors in case a child constructor expects different kind of
arguments.

This commits changes this behavior by requiring a child constructor to
explicitly call its parent constructor if needed.

However the base constructor also had code to ensure proper constructor
instantiation and take care of the stack trace. This code should be
executed regardless the parent constructor is explicitly called or not,
thus the introduction of the SpecialConstructor to handle this: if a
subclass constructor is given, only SpecialConstructor is called before.

Updated the tests and documentation accordingly, and bumped a major.

Closes #6.